+++ src/main/scala/io/viash/functionality/resources/JavaScriptScript.scala
@@ -46,7 +46,7 @@ case class JavaScriptScript(
     val paramsCode = argsAndMeta.map { case (dest, params) =>
     val parSet = params.map { par =>
       // val env_name = par.VIASH_PAR
-      val env_name = Bash.getEscapedArgument(par.VIASH_PAR, "'", """\'""", """\\\'""")
+      val env_name = Bash.getEscapedArgument(par.VIASH_PAR, "String.raw`", "`", """`""", """`+\"`\"+String.raw`""")
 
       val parse = par match {
         case a: BooleanArgumentBase if a.multiple =>