+++ src/main/scala/io/viash/functionality/resources/PythonScript.scala
@@ -46,7 +46,7 @@ case class PythonScript(
     val paramsCode = argsAndMeta.map { case (dest, params) =>
       val parSet = params.map { par =>
       // val env_name = par.VIASH_PAR
-      val env_name = Bash.getEscapedArgument(par.VIASH_PAR, "'", """\'""", """\\\'""")
+      val env_name = Bash.getEscapedArgument(par.VIASH_PAR, "r'", "'", """\'""", """\'\"\'\"r\'""")
 
       val parse = par match {
         case a: BooleanArgumentBase if a.multiple =>