+++ src/main/scala/io/viash/platforms/NextflowVdsl3Platform.scala
@@ -29,6 +29,7 @@ import io.circe.syntax._
 import io.circe.{Printer => JsonPrinter, Json, JsonObject}
 import shapeless.syntax.singleton
 import io.viash.schemas._
+import io.viash.helpers.Escaper
 
 /**
  * Next-gen Platform class for generating NextFlow (DSL2) modules.
@@ -106,8 +107,8 @@ case class NextflowVdsl3Platform(
   // TODO: solve differently
   container: String = "docker"
 ) extends NextflowPlatform {
-  def escapeText(txt: String): String = {
-    Bash.escape(txt, singleQuote = true, newline = true, backtick = false)
+  def escapeSingleQuotedString(txt: String): String = {
+    Escaper(txt, slash = true, singleQuote = true, newline = true)
   }
 
   def modifyFunctionality(config: Config, testing: Boolean): Functionality = {
@@ -155,7 +156,7 @@ case class NextflowVdsl3Platform(
     val versStr = functionality.version.map(ver => s"\n  version = '$ver'").getOrElse("")
 
     val descStr = functionality.description.map{des => 
-      val escDes = escapeText(des)
+      val escDes = escapeSingleQuotedString(des)
       s"\n  description = '$escDes'"
     }.getOrElse("")
 
@@ -163,7 +164,7 @@ case class NextflowVdsl3Platform(
       if (functionality.authors.isEmpty) {
         "" 
       } else {
-        val escAut = escapeText(functionality.authors.mkString(", "))
+        val escAut = escapeSingleQuotedString(functionality.authors.map(_.name).mkString(", "))
         s"\n  author = '$escAut'"
       }
 
@@ -187,7 +188,7 @@ case class NextflowVdsl3Platform(
     
     /************************* HEADER *************************/
     val header = Helper.generateScriptHeader(functionality)
-      .map(h => Bash.escapeMore(h))
+      .map(h => Escaper(h, newline = true))
       .mkString("// ", "\n// ", "")
 
     /************************* SCRIPT *************************/
@@ -202,7 +203,7 @@ case class NextflowVdsl3Platform(
       // if mainResource is a script
       case Some(res) =>
         val code = res.readWithInjection(functionality).get
-        val escapedCode = Bash.escapeMore(code)
+        val escapedCode = Bash.escapeString(code, allowUnescape = true)
           .replace("\\", "\\\\")
           .replace("'''", "\\'\\'\\'")
 
@@ -227,7 +228,7 @@ case class NextflowVdsl3Platform(
       // is memory requirements are defined but directives.memory isn't, use that instead
       memory = directives.memory orElse functionality.requirements.memoryAsBytes.map(_.toString + " B"),
       // is cpu requirements are defined but directives.cpus isn't, use that instead
-      cpus = directives.cpus orElse functionality.requirements.n_proc.map(np => Left(np))
+      cpus = directives.cpus orElse functionality.requirements.cpus.map(np => Left(np))
     )
     val jsonPrinter = JsonPrinter.spaces2.copy(dropNullValues = true)
     val dirJson = directivesToJson.asJson.dropEmptyRecursively()